+++ myems-admin/app/services/settings/combinedequipment/combinedequipmentmeter.service.js
@@ -38,6 +38,21 @@ app.factory('CombinedEquipmentMeterService', function ($http) {
             }, function (response) {
                 callback(response);
             });
+        },
+        
+        // GET combined equipments that a meter is bound to (RESTful style)
+        checkMeterBinding: function(meterId, meterType, headers, callback) {
+            $http.get(getAPI()+'meters/'+meterId+'/combinedequipments', {
+                params: {
+                    type: meterType
+                },
+                headers: headers
+            })
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            });
         }
     };
 });