+++ myems-admin/app/services/settings/shopfloor/shopfloormeter.service.js
@@ -39,6 +39,21 @@ app.factory('ShopfloorMeterService', function($http) {
             }, function (response) {
                 callback(response);
             });
+        },
+        
+        // GET shopfloors that a meter is bound to (RESTful style)
+        checkMeterBinding: function(meterId, meterType, headers, callback) {
+            $http.get(getAPI()+'meters/'+meterId+'/shopfloors', {
+                params: {
+                    type: meterType
+                },
+                headers: headers
+            })
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            });
         }
     };
 });