+++ CHANGELOG.md
@@ -6,11 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added circle check to energy flow diagram in myems-api and myems-admin
+- added duplicate check for binding meter to space in myems-api and myems-admin
+- added duplicate check for binding meter to shopfloor in myems-api and myems-admin
+- added duplicate check for binding meter to combined equipment in myems-api and myems-admin
+- added duplicate check for binding meter to equipment in myems-api and myems-admin
+- added duplicate check for binding meter to store in myems-api and myems-admin
+- added duplicate check for binding meter to tenant in myems-api and myems-admin
+- added myems_production_db.tbl_equipment_hourly to database
+- added modification protection for super administrator in myems-api and myems-admin
 ### Changed
 - changed component import method from static to dynamic imports in myems-web
 - added ErrorBoundary component to capture rendering errors in myems-web
 ### Fixed
 - fixed Overwritten property issue in myems-web
+- fixed datetime picker color theme issue in myems-web
 ### Removed
 - removed print(req.params) from myems-api